debug

+++ Contractive_Autoencoder_in_Pytorch/word_emb.py
@@ -94,7 +94,7 @@ if __name__=="__main__":
 		train_label = train_label[0:60000]
 
 		batch_size = 50
-		batch_num = 60000/batch_size
+		batch_num = int(60000/batch_size)
 		train_data = train_data.reshape((batch_num,batch_size,win_size-1))
 		train_label = train_label.reshape((batch_num,batch_size))
